moved the actual update logic to the PackageService
